+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.15.4 (tbd)
 
 - Fixed issue loading `dotnetjs` in `piral-blazor`
+- Added watcher for *package.json* and config files to `piral debug`
+- Added watcher for *package.json* and config files to `pilet debug`
 - Added `--watch` flag to `piral build` command to `piral-cli`
 - Added `--watch` flag to `pilet build` command to `piral-cli`